2018-01-27 - added: Dumping of DVD which is the parallel track path and dual layer - changed: Visual Studio 2015 to 2017 - fixed: Dumping of audio disc (didn't work from 2017-12-10) - fixed: Rewrite .c2 file if c2 errors exist.           =>  if unfixed errors exist with EdcEcc.exe or edccchk or cdmage though .c2 file rewritten is all zero byte,               maybe the drive misses to report the c2 error.

+++ DiscImageCreator/outputScsiCmdLogforCD.h
@@ -142,6 +142,6 @@ VOID OutputCDSubToLog(
 	LPBYTE lpSubcode,
 	LPBYTE lpSubcodeRaw,
 	INT nLBA,
-	INT byTrackNum,
+	BYTE byTrackNum,
 	FILE* fpParse
 );
